requirements.txt
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
btrfsutil==5.15.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
cupshelpers==1.0
Cython==0.29.26
distlib==0.3.3
distro==1.6.0
dnspython==2.1.0
docopt==0.6.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
keyutils==0.6
lensfun==0.3.95
LibAppArmor==3.0.3
libtorrent===1.2.15-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.15-bindings-python
louis==3.20.0
Markdown==3.3.6
more-itertools==8.10.0
msgpack==1.0.3
netsnmp-python==1.0a1
npyscreen==4.10.5
optimus-manager==1.4
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
pep517==0.12.0
Pillow==8.4.0
pipfile==0.0.2
ply==3.11
progress==1.6
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
PyGObject==3.42.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
pysmbc==1.0.23
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
reportlab==3.6.3
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
six==1.16.0
team==1.0
toml==0.10.2
tomli==1.2.2
udiskie==2.4.0
urllib3==1.26.7
webencodings==0.5.1
youtube-dl==2021.12.17
zipp==3.6.0

Django~=3.2.7
django-autoslug~=1.9.8
django-ckeditor~=6.2.0
django-model-utils~=4.2.0